agentic_rag/requirements.txt
docling
langchain-openai
chromadb
python-dotenv
openai
pydantic
fastapi
uvicorn
python-multipart
transformers
torch
accelerate
pyyaml
trafilatura
gradio
lxml_html_clean
langchain
gitingest
bitsandbytes
ollama
oracledb
sentence-transformers